점수 추가
-좌 상단에 점수 보임

할일:
- 장애물 랜덤소환 말구 파일같은거에서 불러오게..

+++ SlimeRunner/Classes/GameScene.h
@@ -34,7 +34,7 @@ private:
 
     GAME_STATE _gameState = GAME_STATE::PAUSED;
 
-    long long int _score = 0;
+	double _score = 0.0;
 
 };
